+++ lib/index.js
@@ -1,9 +1,13 @@
-// Generated by CoffeeScript 1.9.1
-var EventEmitter, Promise, Waffel, _, async, cheerio, colors, fs, glob, i18n, markdown, marked, matter, md5, moment, nunjucks, path, pushserve, util, yaml,
+// Generated by CoffeeScript 1.10.0
+var EventEmitter, Promise, Waffel, _, async, cheerio, colors, filters, fs, glob, helpers, i18n, markdown, marked, matter, md5, moment, nunjucks, path, pushserve, util, yaml,
   bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; },
   extend = function(child, parent) { for (var key in parent) { if (hasProp.call(parent, key)) child[key] = parent[key]; } function ctor() { this.constructor = child; } ctor.prototype = parent.prototype; child.prototype = new ctor(); child.__super__ = parent.prototype; return child; },
   hasProp = {}.hasOwnProperty,
   indexOf = [].indexOf || function(item) { for (var i = 0, l = this.length; i < l; i++) { if (i in this && this[i] === item) return i; } return -1; };
+
+filters = require('./filters');
+
+helpers = require('./helpers');
 
 _ = require('lodash');
 
@@ -13,7 +17,7 @@ Promise = require('bluebird');
 
 path = require('path');
 
-md5 = require('MD5');
+md5 = require('md5');
 
 EventEmitter = require('events').EventEmitter;
 
@@ -47,6 +51,7 @@ module.exports = Waffel = (function(superClass) {
   extend(Waffel, superClass);
 
   Waffel.prototype.defaults = {
+    silent: false,
     verbose: false,
     defaultPagination: 10,
     defaultSortField: 'slug',
@@ -61,20 +66,19 @@ module.exports = Waffel = (function(superClass) {
     basePath: '',
     assetPath: '',
     root: process.cwd(),
-    ext: '.md',
+    dataExt: '.md',
     templateExt: '.html',
     languages: [],
     defaultLanguage: 'en',
     fallbackLanguage: 'en',
     localiseDefault: false,
     sitemap: true,
+    sitemapName: 'sitemap.xml',
     uglyUrls: false,
     outputExt: '.html',
     displayExt: true,
     dateFormat: 'YYYY-MM-DD',
     markdownOptions: {},
-    helpers: {},
-    filters: {},
     server: false,
     frontmatter: {
       delims: ['---', '---']
@@ -86,191 +90,33 @@ module.exports = Waffel = (function(superClass) {
     }
   };
 
-  Waffel.prototype.helpers = {
-    url: function(name, data, options) {
-      var page, relativeUrl;
-      if (data == null) {
-        data = {};
-      }
-      if (options == null) {
-        options = {};
-      }
-      page = this._getPageByName(name);
-      if (options.page) {
-        page.pagination = {
-          page: options.page
-        };
-      }
-      if (this.options.uglyUrls) {
-        relativeUrl = this._url(page, data, options);
-        if (this.options.displayExt) {
-          return (_.compact([this.options.domain, this.options.basePath, relativeUrl || 'index']).join('/')) + this.options.outputExt;
-        } else {
-          return _.compact([this.options.domain, this.options.basePath, relativeUrl]).join('/');
-        }
-      } else {
-        return _.compact([this.options.domain, this.options.basePath, this._url(page, data, options), 'index.html']).join('/');
-      }
-    },
-    asset: function(_path) {
-      if (_path == null) {
-        _path = '';
-      }
-      return _.compact([this.options.domain, this.options.basePath, this.options.assetPath, _path]).join('/');
-    },
-    absoluteURL: function(url) {
-      return _.compact([this.options.domain, this.options.basePath, url]).join('/');
-    },
-    t: function(key, page) {
-      return i18n.translate(key, {
-        lng: page.language
-      });
-    },
-    loc: function(data, language) {
-      if (data == null) {
-        data = {};
-      }
-      if (language == null) {
-        language = this.options.defaultLanguage;
-      }
-      if (_.isArray(data)) {
-        return data.map((function(_this) {
-          return function(item) {
-            if (item._localised) {
-              return item[language] || item[_this.options.fallbackLanguage];
-            } else {
-              return item;
-            }
-          };
-        })(this));
-      } else if (!data._localised) {
-        return data;
-      } else {
-        return data[language] || data[this.options.fallbackLanguage];
-      }
+  Waffel.prototype.log = function(what) {
+    if (!this.options.silent) {
+      return console.log(what);
     }
   };
 
-  Waffel.prototype.filters = {
-    toArray: function(object) {
-      return _.toArray(object);
-    },
-    pluck: function(object, key) {
-      if (object == null) {
-        object = {};
-      }
-      return _.pluck(object, key);
-    },
-    flatten: function(array) {
-      if (array == null) {
-        array = [];
-      }
-      return _.flatten(array);
-    },
-    uniq: function(array) {
-      if (array == null) {
-        array = [];
-      }
-      return _.uniq(array);
-    },
-    where: function(array, search) {
-      if (array == null) {
-        array = [];
-      }
-      if (search == null) {
-        search = {};
-      }
-      return _.where(array, search);
-    },
-    findWhere: function(array, search) {
-      if (array == null) {
-        array = [];
-      }
-      if (search == null) {
-        search = {};
-      }
-      return _.findWhere(array, search);
-    },
-    limit: function(array, count) {
-      if (array == null) {
-        array = [];
-      }
-      if (count == null) {
-        count = 10;
-      }
-      return array.slice(0, count);
-    },
-    format: function(date, format) {
-      if (format == null) {
-        format = this.options.dateFormat;
-      }
-      return moment(date).format(format);
-    },
-    excerpt: function(text, size) {
-      var $, words;
-      if (size == null) {
-        size = 200;
-      }
-      $ = cheerio.load(marked(text));
-      text = $('p').filter(function(index, element) {
-        return (element.children[0].type === 'text') || _.contains(['em', 'strong'], element.children[0].name);
-      }).first().text().trim();
-      if (text.length > size) {
-        words = text.substring(0, size).split(' ');
-        words.pop();
-        return (words.join(' ')) + "…";
-      } else {
-        return text;
-      }
-    },
-    toJSON: function(data) {
-      return JSON.stringify(data);
-    },
-    inspect: function(object) {
-      console.log(util.inspect(object, false, 2, true));
-      return object;
-    },
-    top: function(data, thresh) {
-      if (thresh == null) {
-        thresh = 3;
-      }
-      data = _.flatten(data);
-      data = _.reduce(data, function(memo, x) {
-        if (memo[x]) {
-          memo[x] = memo[x] + 1;
-        } else {
-          memo[x] = 1;
-        }
-        return memo;
-      }, {});
-      data = _.reduce(data, function(memo, freq, key) {
-        memo.push({
-          key: key,
-          freq: freq
-        });
-        return memo;
-      }, []);
-      data = _.sortBy(data, function(bin) {
-        return -bin.freq;
-      });
-      return data.slice(0, thresh).map(function(x) {
-        return x.key;
-      });
+  Waffel.prototype.error = function(what, e) {
+    console.error(("" + what).red);
+    if (this.options.verbose) {
+      console.error(e);
+    }
+    if (this.options.verbose) {
+      return console.error(e.stack);
     }
   };
 
   function Waffel(opts) {
     this.postGenerate = bind(this.postGenerate, this);
-    var filter, helper, name, ref, ref1, site;
-    this.options = _.extend(this.defaults, opts);
-    this.options.dataFolder = path.join(this.options.root, this.options.dataFolder);
-    this.options.viewFolder = path.join(this.options.root, this.options.viewFolder);
-    this.options.staticFolder = path.join(this.options.root, this.options.staticFolder);
-    this.options.localesFolder = path.join(this.options.root, this.options.localesFolder);
-    this.options.destinationFolder = path.join(this.options.root, this.options.destinationFolder);
-    this.options.structureFile = path.join(this.options.root, this.options.structureFile);
-    this.helpers = _.extend(this.helpers, this.options.helpers);
-    this.filters = _.extend(this.filters, this.options.filters);
+    var e, error, filter, helper, name, ref, ref1, site, structureFileContents;
+    this.options = _.extend({}, this.defaults, opts);
+    ['dataFolder', 'viewFolder', 'staticFolder', 'localesFolder', 'destinationFolder', 'structureFile'].forEach((function(_this) {
+      return function(f) {
+        return _this.options[f] = path.join(_this.options.root, _this.options[f]);
+      };
+    })(this));
+    this.helpers = _.extend({}, helpers, this.options.helpers);
+    this.filters = _.extend({}, filters, this.options.filters);
     ref = this.helpers;
     for (name in ref) {
       helper = ref[name];
@@ -288,7 +134,13 @@ module.exports = Waffel = (function(superClass) {
       return (_.flattenDeep(data.join(''))) + "." + size;
     });
     this.data = {};
-    site = yaml.safeLoad(fs.readFileSync(this.options.structureFile, 'utf8'));
+    try {
+      structureFileContents = fs.readFileSync(this.options.structureFile, 'utf8');
+    } catch (error) {
+      e = error;
+      this.error("Could not locate structureFile: " + this.options.structureFile, e);
+    }
+    site = yaml.safeLoad(structureFileContents);
     this.config = site.config;
     this.structure = site.structure;
     this.registerTemplates();
@@ -314,8 +166,8 @@ module.exports = Waffel = (function(superClass) {
 
   Waffel.prototype.init = function() {
     var _path;
-    _path = path.join(this.options.dataFolder, "**/*" + this.options.ext);
-    console.log("--> Globbing " + _path.cyan + ":");
+    _path = path.join(this.options.dataFolder, "**/*" + this.options.dataExt);
+    this.log("--> Globbing " + _path.cyan + ":");
     i18n.init({
       preload: this.options.languages.concat(['dev']),
       lng: this.options.defaultLanguage,
@@ -335,11 +187,11 @@ module.exports = Waffel = (function(superClass) {
       options = {};
     }
     this.start = process.hrtime();
-    console.log("--> Start generation process...\n---");
+    this.log("--> Start generation process...\n---");
     if (options.data) {
       _.merge(this.data, options.data);
     }
-    this.emit('startGeneration');
+    this.emit('generation:start');
     return fs.ensureDirAsync(this.options.destinationFolder).then((function(_this) {
       return function() {
         var i, language, languages, len, tasks;
@@ -361,7 +213,7 @@ module.exports = Waffel = (function(superClass) {
     var elapsed, millis;
     elapsed = process.hrtime(this.start);
     millis = elapsed[1] / 1000000;
-    console.log("--> Generated " + (pages.length + '').cyan + " pages in " + elapsed[0] + "." + (millis.toFixed(0)) + "s.");
+    this.log("--> Generated " + (pages.length + '').cyan + " pages in " + elapsed[0] + "." + (millis.toFixed(0)) + "s.");
     if (this.options.sitemap) {
       this._createSitemap(pages);
     }
@@ -379,7 +231,7 @@ module.exports = Waffel = (function(superClass) {
       page = ref[name];
       if (page.languages && indexOf.call(page.languages, language) < 0) {
         if (this.options.verbose) {
-          console.log("Notice:".magenta + " " + name.green + " won't be rendered in " + language.yellow);
+          this.log("Notice:".magenta + " " + name.green + " won't be rendered in " + language.yellow);
         }
       } else if (page.template) {
         page.name = name;
@@ -415,7 +267,7 @@ module.exports = Waffel = (function(superClass) {
     });
     data = loadedData.data;
     data.__content = loadedData.content;
-    data.slug = data.slug || path.basename(relativePath, this.options.ext);
+    data.slug = data.slug || path.basename(relativePath, this.options.dataExt);
     if (ref = tokens[1], indexOf.call(this.options.languages, ref) >= 0) {
       language = tokens[1];
       (base1 = this.data[collection])[name1 = data.slug] || (base1[name1] = {
@@ -489,15 +341,28 @@ module.exports = Waffel = (function(superClass) {
   };
 
   Waffel.prototype._renderPage = function(page, _data) {
-    var tmpData;
+    var e, error, tmpData;
     tmpData = {};
     tmpData[page["export"] || 'item'] = _data;
-    return nunjucks.render("" + page.template + this.options.templateExt, _.extend(this.helpers, tmpData, {
-      options: this.options,
-      config: this.config,
-      data: this.data,
-      page: page
-    }));
+    helpers = _.transform(this.helpers, (function(_this) {
+      return function(help, func, key) {
+        return help[key] = function() {
+          [].push.call(arguments, page);
+          return func.apply(_this, arguments);
+        };
+      };
+    })(this));
+    try {
+      return nunjucks.render("" + page.template + this.options.templateExt, _.extend({}, helpers, tmpData, {
+        options: this.options,
+        config: this.config,
+        data: this.data,
+        page: page
+      }));
+    } catch (error) {
+      e = error;
+      return this.log(page.name);
+    }
   };
 
   Waffel.prototype._createCollectionPage = function(page, name, set, language, localised) {
@@ -590,7 +455,7 @@ module.exports = Waffel = (function(superClass) {
         pageInfo = data.slug || data.group || page.group || '';
         pageInfo = pageInfo ? " [" + pageInfo + "]" : '';
         if (_this.options.verbose) {
-          console.log(languageInfo.red + "Generating " + name.green + pageInfo.yellow + paginationInfo.magenta + " at: " + target.cyan);
+          _this.log(languageInfo.red + "Generating " + name.green + pageInfo.yellow + paginationInfo.magenta + " at: " + target.cyan);
         }
         return fs.outputFile(target, output, function(err) {
           return callback(err, {
@@ -605,8 +470,8 @@ module.exports = Waffel = (function(superClass) {
 
   Waffel.prototype._createSitemap = function(pages) {
     var output, target;
-    target = path.join(this.options.destinationFolder, 'sitemap.xml');
-    output = nunjucks.render('sitemap.xml', _.extend(this.helpers, {
+    target = path.join(this.options.destinationFolder, this.options.sitemapName);
+    output = nunjucks.render(this.options.sitemapName, _.extend({}, this.helpers, {
       options: this.options,
       config: this.config,
       data: this.data,
@@ -617,17 +482,17 @@ module.exports = Waffel = (function(superClass) {
     }));
     return fs.outputFile(target, output, (function(_this) {
       return function(err) {
-        return console.log("--> Created " + 'sitemap.xml'.cyan);
+        return _this.log("--> Created " + _this.options.sitemapName.cyan);
       };
     })(this));
   };
 
   Waffel.prototype._launchServer = function() {
     var opts, server;
-    opts = _.extend(this.options.serverConfig, this.options.server);
+    opts = _.extend({}, this.options.serverConfig, this.options.server);
     return server = pushserve(opts, (function(_this) {
       return function() {
-        console.log("--> waffel server waiting for you at " + ("http://localhost:" + opts.port).green);
+        _this.log("--> waffel server waiting for you at " + ("http://localhost:" + opts.port).green);
         return _this.emit('server:start', server);
       };
     })(this));
