+++ package.json
@@ -3,18 +3,19 @@
   "version": "0.0.44",
   "description": "Static website generator, best companion to Vaneella.",
   "main": "lib/index.js",
-  "homepage" : "http://moonwave99.github.io/waffel/",
+  "homepage": "http://moonwave99.github.io/waffel/",
   "author": "moonwave99 <diego.caponera@gmail.com>",
   "repository": {
     "type": "git",
     "url": "git@github.com:moonwave99/waffel.git"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node_modules/.bin/mocha --recursive --reporter spec --timeout 15000 --compilers coffee:coffee-script/register",
+    "prepublish": "coffee --bare -o lib -c src/*.coffee"
   },
   "license": "ISC",
   "dependencies": {
-    "MD5": "^1.2.1",
+    "md5": "^2.0.0",
     "async": "^0.9.0",
     "bluebird": "2.9.27",
     "cheerio": "^0.19.0",
@@ -32,6 +33,9 @@
     "pushserve": "^0.1.6"
   },
   "devDependencies": {
-    "coffee-script": "^1.9.1"
+    "coffee-script": "^1.9.1",
+    "mocha": "^2.3.3",
+    "should": "^7.1.1",
+    "should-promised": "^0.3.1"
   }
 }
